Fix "bad connection" in mysql preset
Many builds failed due to a "bad connection" error that was caused by a
short wait timeout (30 seconds). In environments like github actions
setting up mysql container takes more time, so this option was removed
in favor of default gnomock options (much higher at this moment).

This commit also adds a Ping call to reestablish a connection in case it
was closed by the server.