Return environment aware container address
This is a preparation for running gnomockd from inside a container. For
such future cases, gnomock won't be available to access new containers
through 127.0.0.1 address. Instead, it will need to use host IP address
as seen by the container (it is called Gateway).

Both cases need to be supported since gnomock can run both in a
container and directly on a host. So there is a different Host value
passed to heath and init functions (other usages are from the actual
code running on the host).