Implement lightweight kubernetes (k3s) preset
Implementing this preset made it clear that it's time to change how CI
tests work with swagger specs. Instead of hard coding the SDK version,
which means it must be deployed before the tests can pass, we should
generate a new SDK for each build, and run tests using the latest
versions of everything.