Implement preset generator tool
This is the first attempt to build a CLI utility to generate Gnomock
Presets. It allows to easily generate boilerplate code for new presets
that are about to enter Gnomock ecosystem.

This tool generates preset code itself (should be fully working but
pretty useless), test code, server-related code to support the preset
over HTTP, swagger spec and more.

Of course human intervention is required to build actual presets. For
example, healthcheck function is the bare minimum that needs to be added
to make the preset useful. Initial state setup functions should also be
added wherever possible, but not mandatory.

This tool leaves a bunch of TODO comments wherever users need to make
manual changes.