Implement container cleanup mechanism
From now, every container started using gnomock will have a sidecar
container whose only purpose is to make sure it is cleaned up when the
tests complete.

To trigger the cleanup, send a request to /sync/:id. It will hang (make
sure timeout is long enough). When tests complete, the program exits,
and the request it released automatically, signaling the sidecar that
it's time. After that, regardless of the success of container
termination, cleaner process stops.

If a user chooses to use `gnomock.Stop()` anyway, the sidecar will be
terminated immediately. For that, sidecar id is now encoded into
container id as returned to the user (both directly and over HTTP)