Use longer timeouts in tests
For some reason, with a recent docker update (docker for desktop 3.3.1
on mac) tests started to fail with timeouts. I'm not sure what the
reason for this slowness is, but giving the tests more time to complete
fixes the symptoms.

While visually inspecting a list of running containers during tests I
noticed that it takes much longer now to create a single container.
Maybe this is related.