Support docker registry auth for private images
This commit adds WithRegistryAuth option that allows to access private
docker images, which were not supported before. This feature is not
going to be a part of CI process, but I tested it once with valid and
invalid credentials to access a private docker image, and it looked
good.

There might be a way to "guess" user's credentials as they are likely
stored in ~/.docker/config.json file. On mac, I don't see the creds in
the file, but in Ubuntu it is probably different. For now this should be
good enough for most users.