Support explicit port mappings in remote mode
When gnomock runs in server mode, it is impossible to control the port
mappings created by the presets. In this commit I add an ability to
override preset-specific named ports by any other values in order to
control which ports are allocated by the remote gnomock instance.

This feature is a bit fragile, because the user is required to study the
preset before attempting to override its built-in values. Custom named
ports must expose the ports with the same names and internal ports for
the override to work. Basically, only the "host port" value should be
changed.

This commit fixes https://github.com/orlangure/gnomock/issues/441.