Allow users to set number of partitions in topic
I have added new type, TopicConfig, and added it to options WithTopicConfig
in backwards-compatible way

For now it just has topic name and number of partitions, but it can be enhanced in the future with other options.

I wanted to also add a replication count, but there is just one broker in the cluster, so it makes no sense.